requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.7.0
python-dotenv==1.1.0
PyJWT==2.10.1
psycopg2-binary==2.9.9
